V3.8.3:优化逻辑,优先从otatools.zip中获取host端做包工具
修改点:1 优化逻辑,优先从otatools.zip中获取host端做包工具
测试建议:可以完成OTA编译

+++ tools/init.sh
@@ -112,13 +112,27 @@ printf "%s\n" "=========================所有信息END=========================
 
 
 printf "=================将host端工具从out拷贝到%s/linux-x86下======================\n" $otabuild
-if [ ! -d $otabuild/linux-x86 ]; then mkdir -p $otabuild/linux-x86; fi
-if [ ! -d $otabuild/linux-x86/bin ]; then mkdir $otabuild/linux-x86/bin; fi
-if [ ! -d $otabuild/linux-x86/framework ]; then mkdir $otabuild/linux-x86/framework; fi
-if [ ! -d $otabuild/linux-x86/lib64 ]; then mkdir $otabuild/linux-x86/lib64; fi
-cp -vu $ANDROID/out/host/linux-x86/bin/bsdiff $ANDROID/out/host/linux-x86/bin/imgdiff $otabuild/linux-x86/bin/
-cp -vu $ANDROID/out/host/linux-x86/framework/signapk.jar $otabuild/linux-x86/framework/
-cp -vu $ANDROID/out/host/linux-x86/lib64/libc++.so $ANDROID/out/host/linux-x86/lib64/libconscrypt_openjdk_jni.so $ANDROID/out/host/linux-x86/lib64/libdivsufsort.so $ANDROID/out/host/linux-x86/lib64/libdivsufsort64.so $otabuild/linux-x86/lib64/
+if [ -e $ANDROID/out/dist/otatools.zip ]; then
+  # if otatools.zip exits, always extract it to get all of host building tools firstly. otatools.zip is used for this purpose.
+  unzip -o $ANDROID/out/dist/otatools.zip "bin/*" "framework/*" "lib64/*" -d $otabuild/linux-x86/
+else
+  # if otatools.zip didn't exit, copy essential file manually.
+  if [ ! -d $otabuild/linux-x86 ]; then mkdir -p $otabuild/linux-x86; fi
+  if [ ! -d $otabuild/linux-x86/bin ]; then mkdir $otabuild/linux-x86/bin; fi
+  if [ ! -d $otabuild/linux-x86/framework ]; then mkdir $otabuild/linux-x86/framework; fi
+  if [ ! -d $otabuild/linux-x86/lib64 ]; then mkdir $otabuild/linux-x86/lib64; fi
+  cp -vu $ANDROID/out/host/linux-x86/bin/bsdiff                                                                    $otabuild/linux-x86/bin/
+  cp -vu $ANDROID/out/host/linux-x86/bin/imgdiff                                                                   $otabuild/linux-x86/bin/
+  cp -vu $ANDROID/out/host/linux-x86/framework/signapk.jar                                                         $otabuild/linux-x86/framework/
+  cp -vu $ANDROID/out/host/linux-x86/lib64/libc++.so                                                               $otabuild/linux-x86/lib64/
+  cp -vu $ANDROID/out/host/linux-x86/lib64/libconscrypt_openjdk_jni.so                                             $otabuild/linux-x86/lib64/
+  if [ $BIGVERSION -lt 8 ]; then
+    cp -vu $ANDROID/out/host/linux-x86/lib64/libdivsufsort.so $ANDROID/out/host/linux-x86/lib64/libdivsufsort64.so $otabuild/linux-x86/lib64/
+  elif [ $BIGVERSION -ge 8 ]; then
+    cp -vu $ANDROID/out/host/linux-x86/bin/bro                                                                     $otabuild/linux-x86/bin/
+    cp -vu $ANDROID/out/host/linux-x86/lib64/libbrotli.so                                                          $otabuild/linux-x86/lib64/
+  fi
+fi
 
 printf "=================将target-files从/mnt/hgfs拷贝到%s/input下======================\n" $otabuild
 cp -vf $target_old_win $target_old_dir
